Create a command for setting speed for the drivetrain
