--- src/main/java/robot/subsystems/Commands/Examplecommand.java
@@ -1,31 +1,0 @@
-package robot.subsystems.Commands;
-
-import edu.wpi.first.wpilibj.command.Command;
-
-public class Examplecommand extends Command {
-
-    @Override
-    protected void initialize() {
-
-    }
-
-    @Override
-    protected void execute() {
-
-    }
-
-    @Override
-    protected boolean isFinished() {
-        return false;
-    }
-
-    @Override
-    protected void interrupted() {
-
-    }
-
-    @Override
-    protected void end() {
-
-    }
-}
